Fixed compile against Qt < 5.4
Affects mainly Ubuntu 14.04 (Qt 5.2) and 14.10 (Qt 5.3).

+++ src/tiled/tilestamp.cpp
@@ -299,7 +299,7 @@ TileStamp TileStamp::fromJson(const QJsonObject &json, const QDir &mapDir)
     TileStamp stamp;
 
     stamp.setName(json.value(QLatin1String("name")).toString());
-    stamp.setQuickStampIndex(json.value(QLatin1String("quickStampIndex")).toInt(-1));
+    stamp.setQuickStampIndex(static_cast<int>(json.value(QLatin1String("quickStampIndex")).toDouble(-1)));
 
     QJsonArray variations = json.value(QLatin1String("variations")).toArray();
     for (const QJsonValue &value : variations) {
